refactor: reorganize project into workspaces
- Split the project into multiple workspaces to improve modularity and maintainability.
- Each workspace focuses on a specific component of the project.
- Updated `Cargo.toml` files to reflect the new structure.
