+++ urdekcah/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "urdekcah"
+authors.workspace = true
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+tokio.workspace = true
+anyhow.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
+weather.workspace = true
