+++ weather/src/error.rs
