+++ weather/src/models/api.rs
