+++ weather/src/models/weather.rs
