+++ weather/src/service/mod.rs
