+++ weather/src/service/weather.rs
